# frontend/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetailResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": ["detail"],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "WorkspaceMemoryActionRequest": {
        "properties": {
          "actor": {
            "default": "operator",
            "title": "Actor",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "WorkspaceMemoryActionRequest",
        "type": "object"
      },
      "WorkspaceMemoryAddRequest": {
        "properties": {
          "confirmed_by": {
            "default": "operator",
            "title": "Confirmed By",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "source_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Label"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": ["session_id", "kind", "content"],
        "title": "WorkspaceMemoryAddRequest",
        "type": "object"
      },
      "WorkspaceMemoryCandidateDecisionRequest": {
        "properties": {
          "actor": {
            "default": "operator",
            "title": "Actor",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": ["session_id"],
        "title": "WorkspaceMemoryCandidateDecisionRequest",
        "type": "object"
      },
      "WorkspaceMemoryCandidateListPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceMemoryCandidateResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfoResponse"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": ["session_id", "page", "items"],
        "title": "WorkspaceMemoryCandidateListPageResponse",
        "type": "object"
      },
      "WorkspaceMemoryCandidateRejectedResponse": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/WorkspaceMemoryCandidateResponse"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "rejected_by": {
            "title": "Rejected By",
            "type": "string"
          }
        },
        "required": ["candidate", "rejected_by", "reason"],
        "title": "WorkspaceMemoryCandidateRejectedResponse",
        "type": "object"
      },
...
      "WorkspaceMemoryDetailResponse": {
        "properties": {
          "entry": {
            "$ref": "#/components/schemas/WorkspaceMemoryEntryResponse"
          }
        },
        "required": ["entry"],
        "title": "WorkspaceMemoryDetailResponse",
        "type": "object"
      },
...
      "WorkspaceMemoryKind": {
        "description": "Operator-facing categories for durable workspace memory.",
        "enum": [
          "fact",
          "convention",
          "command",
          "failure_pattern",
          "architecture_note",
          "user_preference",
          "task_outcome"
        ],
        "title": "WorkspaceMemoryKind",
        "type": "string"
      },
      "WorkspaceMemoryListPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceMemoryEntryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfoResponse"
          }
        },
        "required": ["page", "items"],
        "title": "WorkspaceMemoryListPageResponse",
        "type": "object"
      },
...
      "WorkspaceMemoryPrunePreviewResponse": {
        "properties": {
          "entry": {
            "$ref": "#/components/schemas/WorkspaceMemoryEntryResponse"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "would_prune": {
            "title": "Would Prune",
            "type": "boolean"
          }
        },
        "required": ["entry", "would_prune"],
        "title": "WorkspaceMemoryPrunePreviewResponse",
        "type": "object"
      },
      "WorkspaceMemoryState": {
        "description": "Lifecycle states for workspace-scoped memory entries.",
        "enum": ["active", "stale", "invalidated", "imported", "pruned"],
        "title": "WorkspaceMemoryState",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Glassbox",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/memory": {
      "get": {
        "description": "Return a bounded page of projected workspace memory entries.",
        "operationId": "list_workspace_memory_page_memory_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WorkspaceMemoryState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/WorkspaceMemoryKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "include_pruned",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Pruned",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryListPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Workspace Memory Page"
      },
      "post": {
        "description": "Create and immediately confirm an operator-provided memory entry.",
        "operationId": "add_workspace_memory_memory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceMemoryAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Workspace Memory"
      }
    },
    "/memory/candidates": {
      "get": {
        "description": "Return operator-reviewable memory candidates for one session.",
        "operationId": "list_workspace_memory_candidates_memory_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryCandidateListPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Workspace Memory Candidates"
      }
    },
    "/memory/candidates/{candidate_id}/confirm": {
      "post": {
        "description": "Create confirmed memory from one generated candidate.",
        "operationId": "confirm_workspace_memory_candidate_memory_candidates__candidate_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceMemoryCandidateDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Workspace Memory Candidate"
      }
    },
    "/memory/candidates/{candidate_id}/reject": {
      "post": {
        "description": "Record explicit operator rejection for one generated candidate.",
        "operationId": "reject_workspace_memory_candidate_memory_candidates__candidate_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceMemoryCandidateDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryCandidateRejectedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Workspace Memory Candidate"
      }
    },
    "/memory/{memory_id}": {
      "get": {
        "description": "Return one projected workspace memory entry.",
        "operationId": "get_workspace_memory_detail_memory__memory_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workspace Memory Detail"
      }
    },
    "/memory/{memory_id}/confirm": {
      "post": {
        "description": "Confirm or refresh confirmation evidence for one memory entry.",
        "operationId": "confirm_workspace_memory_memory__memory_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceMemoryActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Workspace Memory"
      }
    },
    "/memory/{memory_id}/invalidate": {
      "post": {
        "description": "Invalidate one memory entry while keeping it inspectable.",
        "operationId": "invalidate_workspace_memory_memory__memory_id__invalidate_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceMemoryActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invalidate Workspace Memory"
      }
    },
    "/memory/{memory_id}/prune": {
      "post": {
        "description": "Prune one memory entry from active retrieval while preserving history.",
        "operationId": "prune_workspace_memory_memory__memory_id__prune_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceMemoryActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prune Workspace Memory"
      }
    },
    "/memory/{memory_id}/prune-preview": {
      "post": {
        "description": "Preview a prune action without mutating canonical memory events.",
        "operationId": "preview_workspace_memory_prune_memory__memory_id__prune_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceMemoryActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMemoryPrunePreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetailResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Workspace Memory Prune"
      }
    },
...
  }
}